fixed "comparison between signed & unsigned int" warning on line 117, in processRGBAImage()
git-svn-id: https://pd-gem.svn.sourceforge.net/svnroot/pd-gem/trunk/Gem@599 49b2a0ca-c541-4d49-a7ef-d785e4371512

+++ src/Pixes/pix_lumaoffset.cpp
@@ -114,7 +114,7 @@ void pix_lumaoffset :: processRGBAImage(imageStruct &image)
 	    pCurrentOutput+=(nWidth*nLineGap);
 	}
     } else {
-      if (hPreviousLineHeights==NULL || hPreviousLineHeights_size<nWidth*sizeof(U32*)){
+      if (hPreviousLineHeights==NULL || hPreviousLineHeights_size<(int)(nWidth*sizeof(U32*)) ){
 	free(hPreviousLineHeights);
 	hPreviousLineHeights_size=nWidth*sizeof(U32*);
 	hPreviousLineHeights=malloc(hPreviousLineHeights_size);
